wifi: Increase kMaxStopCompleteWaitMs to 250 msec.
The current timeout value of 100msec for kMaxStopCompleteWaitMs
is not sufficient to stop the Wi-Fi HAL. QC Wi-Fi HAL also needs
to stop the event thread by sending an exit msg as part of stop WiFi HAL.
This stop operations are not completed with in the stipulated 100msec.

Hence , set this timeout value to 250msec.

CRs-Fixed: 2357953
Change-Id: I8680aea7d143dc4676a14654e5fbaf77650588bb